requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
openai
python-dotenv

loguru
pytest
httpx
pytest-mock